+++ pystapi-client/src/pystapi_client/stapi_api_io.py
@@ -147,7 +147,7 @@ class StapiIO:
             resp = self.session.send(modified)
         except Exception as err:
             logger.debug(err)
-            raise APIError.from_response(resp)
+            raise APIError(f"Error sending request: {err}")
 
         # NOTE what about other successful status codes?
         if resp.status_code != 200: